# openapi.yaml
openapi: 3.1.0
info:
  title: PPshkin API
  version: 1.0.0
  description: HTTP API мини-приложения ППшкин в MAX. Гость ведёт дневник питания
    и получает персональные предложения заведений, заведение управляет меню,
    горящими позициями и бронями.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Сессионный токен из POST /api/v1/auth/max или тестовый токен
        демо-учётки (только при DEMO_MODE=true).
  schemas:
...
    ManualMealInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          description: Название блюда
        kcal:
          type: integer
          minimum: 0
          maximum: 5000
          description: Калорийность одним числом, вместо пары kcalMin и kcalMax
        kcalMin:
          type: integer
          minimum: 0
          maximum: 5000
          description: Нижняя граница калорийности, передаётся вместе с kcalMax
        kcalMax:
          type: integer
          minimum: 0
          maximum: 5000
          description: Верхняя граница калорийности, не меньше kcalMin
        proteinG:
          description: Белки, г
          type: number
          minimum: 0
          maximum: 500
        fatG:
          description: Жиры, г
          type: number
          minimum: 0
          maximum: 500
        carbsG:
          description: Углеводы, г
          type: number
          minimum: 0
          maximum: 500
        tags:
          maxItems: 10
          type: array
          items:
            type: string
            enum:
              - sweet
              - dessert
              - pastry
              - chocolate
              - fruit
              - berries
              - dairy
              - cheese
              - eggs
              - meat
              - poultry
              - fish
              - seafood
              - vegetarian
              - vegetables
              - greens
              - salad
              - soup
              - grain
              - pasta
              - rice
              - potato
              - bread
              - fried
              - grilled
              - spicy
              - fast_food
              - coffee
              - tea
              - drink
              - juice
              - high_protein
              - light
              - hearty
              - breakfast
        eatenAt:
          type: string
          format: date-time
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
          description: "Когда съедено, ISO 8601: не раньше чем 7 дней назад и не позже чем
            через 5 минут от текущего времени"
      required:
        - title
      description: "Приём пищи, введённый вручную. Калорийность: либо kcal, либо пара
        kcalMin и kcalMax. Без eatenAt запись получает текущее время"
...
    Problem:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        code:
          type: string
          description: Машиночитаемый код ошибки
        detail:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
              message:
                type: string
            required:
              - path
              - message
            additionalProperties: false
      required:
        - type
        - title
        - status
        - code
        - detail
      additionalProperties: false
      description: Описание ошибки по RFC 9457
...
    Meal:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        kcalMin:
          type: integer
        kcalMax:
          type: integer
        kcal:
          type: integer
          description: Середина диапазона, по ней считаются итоги дня
        proteinG:
          description: Белки, г
          type:
            - number
            - "null"
        fatG:
          description: Жиры, г
          type:
            - number
            - "null"
        carbsG:
          description: Углеводы, г
          type:
            - number
            - "null"
        tags:
          type: array
          items:
            type: string
            enum:
              - sweet
              - dessert
              - pastry
              - chocolate
              - fruit
              - berries
              - dairy
              - cheese
              - eggs
              - meat
              - poultry
              - fish
              - seafood
              - vegetarian
              - vegetables
              - greens
              - salad
              - soup
              - grain
              - pasta
              - rice
              - potato
              - bread
              - fried
              - grilled
              - spicy
              - fast_food
              - coffee
              - tea
              - drink
              - juice
              - high_protein
              - light
              - hearty
              - breakfast
        source:
          type: string
          enum:
            - photo
            - text
            - manual
            - booking
            - demo
          description: photo и text - распознано, manual - введено вручную, booking -
            погашенная бронь, demo - демо-данные
        confidence:
          anyOf:
            - type: number
              minimum: 0
              maximum: 1
            - type: "null"
          description: Уверенность распознавания от 0 до 1, null для остальных записей
        eatenAt:
          type: string
          format: date-time
          examples:
            - 2026-09-25T12:00:00.000Z
          description: Дата и время в формате ISO 8601, UTC
        slot:
          type: string
          enum:
            - breakfast
            - lunch
            - snack
            - dinner
          description: Приём пищи по местному времени пользователя
      required:
        - id
        - title
        - kcalMin
        - kcalMax
        - kcal
        - proteinG
        - fatG
        - carbsG
        - tags
        - source
        - confidence
        - eatenAt
        - slot
      additionalProperties: false
      description: Запись дневника питания. Калорийность приблизительная
    DayTotals:
      type: object
      properties:
        meals:
          type: integer
          description: Количество записей
        kcalMin:
          type: integer
        kcalMax:
          type: integer
        kcal:
          type: integer
          description: Сумма середин диапазонов
        proteinG:
          type: number
        fatG:
          type: number
        carbsG:
          type: number
      required:
        - meals
        - kcalMin
        - kcalMax
        - kcal
        - proteinG
        - fatG
        - carbsG
      additionalProperties: false
      description: Итоги дня
    DiaryDay:
      type: object
      properties:
        date:
          type: string
          format: date
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
          description: Календарная дата YYYY-MM-DD в часовом поясе пользователя
        timezone:
          type: string
          description: Часовой пояс, в котором считаются границы дня
        targetKcal:
          type: integer
          description: Ориентир калорийности пользователя
        totals:
          $ref: "#/components/schemas/DayTotals"
        remainingKcal:
          type: integer
          minimum: 0
          description: Сколько осталось до ориентира, не меньше 0
        meals:
          type: array
          items:
            $ref: "#/components/schemas/Meal"
          description: Записи дня по возрастанию eatenAt
      required:
        - date
        - timezone
        - targetKcal
        - totals
        - remainingKcal
        - meals
      additionalProperties: false
      description: Дневник за один день
    DiarySummaryDay:
      type: object
      properties:
        date:
          type: string
          format: date
          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))$
          description: Календарная дата YYYY-MM-DD в часовом поясе пользователя
        meals:
          type: integer
        kcal:
          type: integer
        kcalMin:
          type: integer
        kcalMax:
          type: integer
        proteinG:
          type: number
        fatG:
          type: number
        carbsG:
          type: number
      required:
        - date
        - meals
        - kcal
        - kcalMin
        - kcalMax
        - proteinG
        - fatG
        - carbsG
      additionalProperties: false
      description: Итоги одного дня, дни без записей с нулями
    DiarySummary:
      type: object
      properties:
        days:
          type: array
          items:
            $ref: "#/components/schemas/DiarySummaryDay"
          description: Дни от старых к новым, последний - сегодня
      required:
        - days
      additionalProperties: false
      description: Итоги по дням
...
paths:
...
  /api/v1/diary/today:
    get:
      operationId: getDiaryToday
      summary: Дневник за сегодня
      tags:
        - diary
      description: "Сегодняшний день считается в часовом поясе пользователя. Коды
        ошибок: user_not_found (404): аккаунт удалён, войдите заново."
      security: *a3
      responses:
        "200":
          description: Дневник за сегодня
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryDay"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/diary/days/{date}:
    get:
      operationId: getDiaryDay
      summary: Дневник за день
      tags:
        - diary
      description: "Границы дня считаются в часовом поясе пользователя. Коды ошибок:
        validation_failed (400): передайте дату в формате YYYY-MM-DD.
        user_not_found (404): аккаунт удалён, войдите заново."
      parameters:
        - schema:
            type: string
            format: date
          example: 2026-09-25
          in: path
          name: date
          required: true
          description: Дата YYYY-MM-DD в часовом поясе пользователя
      security: *a3
      responses:
        "200":
          description: Дневник за день
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiaryDay"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/diary/summary:
    get:
      operationId: getDiarySummary
      summary: Итоги по дням
      tags:
        - diary
      description: "Итоги за последние days дней включая сегодня, от старых к новым.
        Дни без записей тоже есть в ответе, с нулями. Коды ошибок:
        validation_failed (400): days от 1 до 31. user_not_found (404): аккаунт
        удалён, войдите заново."
      parameters:
        - schema:
            default: 7
            type: integer
            minimum: 1
            maximum: 31
          in: query
          name: days
          required: false
          description: Сколько последних дней включая сегодня
      security: *a3
      responses:
        "200":
          description: Итоги по дням
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiarySummary"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/diary/meals:
    post:
      operationId: createMeal
      summary: Записать приём пищи вручную
      tags:
        - diary
      description: "Ручной ввод и подтверждение кандидата из распознавания. Коды
        ошибок: validation_failed (400): исправьте поля, перечисленные в errors.
        consent_required (403): нет согласия на обработку данных, покажите его
        текст из GET /api/v1/consents и отправьте PUT
        /api/v1/consents/personal_data. user_not_found (404): аккаунт удалён,
        войдите заново. eaten_at_out_of_range (422): время приёма пищи должно
        быть не раньше чем 7 дней назад и не позже чем через 5 минут."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ManualMealInput"
      security: *a3
      responses:
        "201":
          description: Запись создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: Unprocessable Entity
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/diary/meals/photo:
    post:
      operationId: logMealFromPhoto
      summary: Записать приём пищи по фото
      tags:
        - diary
      description: "Фото передаётся в поле image формы multipart/form-data: JPEG, PNG
        или WebP до 10 МБ, тип определяется по содержимому. Фото не сохраняется
        и уходит в распознавание без данных пользователя. Распознавание платное,
        поэтому не больше 20 запросов в минуту. Уверенно распознанные блюда
        сохраняются сразу, неуверенные возвращаются кандидатами: гость
        подтверждает их через POST /api/v1/diary/meals. При status unavailable
        предложите ручной ввод. Коды ошибок: image_required, image_empty,
        invalid_multipart (400): приложите одно фото в поле image.
        image_too_large, upload_too_many_parts (413): уменьшите фото и отправьте
        только его. multipart_required, unsupported_image_type (415): отправьте
        JPEG, PNG или WebP как multipart/form-data. consent_required (403): нет
        согласия на обработку данных, покажите его текст из GET /api/v1/consents
        и отправьте PUT /api/v1/consents/personal_data. user_not_found (404):
        аккаунт удалён, войдите заново."
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - image
              properties:
                image:
                  type: string
                  format: binary
                  description: Фото JPEG, PNG или WebP до 10 МБ
      security: *a3
      responses:
        "200":
          description: Результат распознавания
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealLogResult"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "413":
          description: Payload Too Large
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "415":
          description: Unsupported Media Type
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/diary/meals/text:
    post:
      operationId: logMealFromText
      summary: Записать приём пищи по описанию
      tags:
        - diary
      description: "Распознавание платное, поэтому не больше 20 запросов в минуту.
        Уверенно распознанные блюда сохраняются сразу, неуверенные возвращаются
        кандидатами: гость подтверждает их через POST /api/v1/diary/meals. При
        status unavailable предложите ручной ввод. Коды ошибок:
        validation_failed (400): исправьте поля, перечисленные в errors.
        consent_required (403): нет согласия на обработку данных, покажите его
        текст из GET /api/v1/consents и отправьте PUT
        /api/v1/consents/personal_data. user_not_found (404): аккаунт удалён,
        войдите заново."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  minLength: 1
                  maxLength: 500
                  description: Что съедено, своими словами, например «овсянка с бананом и кофе с
                    молоком»
              required:
                - description
      security: *a3
      responses:
        "200":
          description: Результат распознавания
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MealLogResult"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /api/v1/diary/meals/{id}:
    patch:
      operationId: updateMeal
      summary: Исправить запись
      tags:
        - diary
      description: "Меняет только переданные поля. Коды ошибок: validation_failed
        (400): исправьте поля, перечисленные в errors. consent_required (403):
        нет согласия на обработку данных, покажите его текст из GET
        /api/v1/consents и отправьте PUT /api/v1/consents/personal_data.
        meal_not_found (404): записи нет или она принадлежит другому
        пользователю. user_not_found (404): аккаунт удалён, войдите заново.
        eaten_at_out_of_range (422): время приёма пищи должно быть не раньше чем
        7 дней назад и не позже чем через 5 минут."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MealPatchInput"
      parameters:
        - schema:
            type: integer
            exclusiveMinimum: 0
          in: path
          name: id
          required: true
      security: *a3
      responses:
        "200":
          description: Исправленная запись
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Meal"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "403":
          description: Forbidden
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "422":
          description: Unprocessable Entity
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    delete:
      operationId: deleteMeal
      summary: Удалить запись
      tags:
        - diary
      description: "Работает без согласия на обработку данных. Коды ошибок:
        validation_failed (400): id должен быть положительным целым.
        meal_not_found (404): записи нет или она принадлежит другому
        пользователю."
      parameters:
        - schema:
            type: integer
            exclusiveMinimum: 0
          in: path
          name: id
          required: true
      security: *a3
      responses:
        "204":
          description: Запись удалена
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "401":
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a2
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "429":
          description: Too Many Requests
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          headers: *a1
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
servers:
  - url: /
    description: Текущий хост
...